requirements.txt
APScheduler==3.6.3
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
flask-requests==0.0.14
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pampy==0.3.0
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
